Fix a problem with tablebases with larger numbers of namesblocks
The newly created larger array was not swapped with the old array.
This caused lots of invalid array accesses and, sometimes, an fseek
to an undefined location resulting in incorrectly sized output files.
See Issue #318 .

Also when comparing contents the block under consideration was not
correctly updated. This caused valgrind to report an Illegal write.
I have not found any actual runtime issues caused by this.

Could this be the cause of an old issue from the forum, where I
described how FORM writes a really enormous amount of data to disk
when creating relatively small tablebase files? For this reason it
is a very bad idea to create a tablebase directly in a network
directory.